+++ plugins/nf-quilt/src/main/nextflow/quilt/QuiltObserver.groovy
@@ -22,7 +22,7 @@ import nextflow.quilt.nio.QuiltPathFactory
 import nextflow.trace.TraceObserver
 
 import java.nio.file.Path
-// import java.util.concurrent.locks.Lock
+import java.util.concurrent.locks.Lock
 import java.util.concurrent.locks.ReentrantLock
 
 import groovy.transform.CompileStatic
@@ -38,12 +38,13 @@ import groovy.util.logging.Slf4j
 class QuiltObserver implements TraceObserver {
 
     private Session session
+    private String workDir
 
     // Is this overkill? Do we only ever have one output URI per run?
     private String[] outputPrefixes = ['pub', 'out']
     final private Map<String,String> outputURIs = [:]
     final private Map<String, Map<String, Path>> packageOverlays = [:]
-    // final private Lock lock = new ReentrantLock() // Need this because of threads
+    final private Lock lock = new ReentrantLock() // Need this because of threads
 
     static QuiltPath asQuiltPath(Path path) {
         if (path in QuiltPath) {
@@ -57,15 +58,11 @@ class QuiltObserver implements TraceObserver {
         return null
     }
 
-    static String pkgKey(QuiltPath path) {
-        return "${path.getBucket()}/${path.getPackageName()}"
-    }
-
     static String quiltURIfromS3(String s3uri) {
-        println("quiltURIfromS3: $s3uri")
+        log.debug("quiltURIfromS3: $s3uri")
         String[] partsArray = s3uri.split('/')
         List<String> parts = new ArrayList(partsArray.toList())
-        parts.eachWithIndex { p, i -> println("quiltURIfromS3.parts[$i]: $p") }
+        // parts.eachWithIndex { p, i -> println("quiltURIfromS3.parts[$i]: $p") }
 
         if (parts.size() < 2) {
             throw new IllegalArgumentException("Invalid s3uri[${parts.size()}]: $parts")
@@ -76,14 +73,15 @@ class QuiltObserver implements TraceObserver {
         }
         String bucket = parts.remove(0)
         String dest = parts.join('%2f')
-        println("quiltURIfromS3.bucket: $bucket dest: $dest from $parts")
         String suffix = parts.size() > 1 ? parts.removeLast() : 'default_suffix'
-        println("quiltURIfromS3.suffix: $suffix from $parts")
         String prefix = parts.size() > 0 ? parts.removeLast() : 'default_prefix'
-        println("quiltURIfromS3.prefix: $prefix from $parts")
         String base = "quilt+s3://${bucket}#package=${prefix}%2f${suffix}"
         String uri = base + '&dest=' + ((dest) ?: '/')
         return uri
+    }
+
+    static String pkgKey(QuiltPath path) {
+        return "${path.getBucket()}/${path.getPackageName()}"
     }
 
     void findOutputParams(Map<String, Object> params) {
@@ -116,26 +114,56 @@ class QuiltObserver implements TraceObserver {
         }
     }
 
-    boolean confirmPath(QuiltPath path) {
-        log.debug("confirmPath[$path]")
-        String key = pkgKey(path)
-        if (!outputURIs.containsKey(key)) {
-            log.warn("Output URI not found for key[$key] from path[$path]")
-            return false
+    String workRelative(Path src) {
+        Path source = src.toAbsolutePath().normalize()
+        log.debug("workRelative.source: $source [type: ${source.getClass()}]")
+        Path workDir = session.workDir.toAbsolutePath().normalize()
+        log.debug("workRelative.workDir: $workDir [type: ${workDir.getClass()}]")
+        try {
+            Path subPath = workDir.relativize(source)
+            log.debug("workRelative.subPath: $subPath]")
+            // drop first two components, which are the workDir
+            Path relPath = subPath.subpath(2, subPath.getNameCount())
+            log.debug("workRelative.relPath: $relPath]")
+            return relPath.toString()
+        } catch (IllegalArgumentException e) {
+            log.error("workRelative.fallback: $e")
+            log.warn("Cannot relativize source:${source.getClass()} to workDir:${workDir.getClass()}")
+            return source.toString()
         }
-        return true
     }
 
-    boolean matchPath(String path) {
-        log.debug("matchPath[$path]")
+    void addOverlay(String pkgKey, Path source) {
+        lock.lock()
+        try {
+            Map<String, Path> overlays = packageOverlays.get(pkgKey, [:]) as Map<String, Path>
+            String relPath = workRelative(source)
+            overlays[relPath] = source
+            packageOverlays[pkgKey] = overlays
+        } finally {
+            lock.unlock()
+        }
+    }
+
+    boolean confirmQuiltPath(QuiltPath qPath) {
+        log.debug("confirmQuiltPath[$qPath]")
+        String key = pkgKey(qPath)
+        return outputURIs.containsKey(key) ? true : false
+    }
+
+    boolean canOverlayPath(Path dest, Path source) {
+        log.debug("canOverlayPath[$dest] <- $source")
         Set<String> keys = outputURIs.keySet()
+        log.debug("canOverlayPath: keys[${keys.size()}]: $keys")
         for (String key : keys) {
-            if (path.contains(key)) {
-                log.debug("matchPath: matched key[$key] to $path")
+            log.debug("canOverlayPath: checking key[$key] for $dest")
+            if (dest.toString().contains(key)) {
+                log.warn("canOverlayPath: matched key[$key] to $dest")
+                addOverlay(key, source)
                 return true
             }
         }
-        log.warn("matchPath: no key found for $path in $keys")
+        log.error("canOverlayPath: no key found for $dest in $keys")
         return false
     }
 
@@ -143,6 +171,7 @@ class QuiltObserver implements TraceObserver {
     void onFlowCreate(Session session) {
         log.debug("`onFlowCreate` $this")
         this.session = session
+        this.workDir = session.config.workDir
         findOutputParams(session.getParams())
         checkConfig(session.config)
     }
@@ -153,10 +182,9 @@ class QuiltObserver implements TraceObserver {
         // Path source may be null, won't work with older versions of Nextflow
         log.debug("onFilePublish.Path[$destination] <- $source")
         QuiltPath qPath = asQuiltPath(destination)
-        if (qPath != null) {
-            confirmPath(qPath)
-        } else {
-            matchPath(destination.toString())
+        boolean ok = (qPath != null) ? confirmQuiltPath(qPath) : canOverlayPath(destination, source)
+        if (!ok) {
+            log.error("onFilePublish: no match for $destination")
         }
     }