+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.8.2] 2024-09-05
 
-- Use copyFile rather than writeString for overlay files
+- Use copyFile rather than writeString for overlay files [requires NextFlow 23 or later]
 
 ## [0.8.1] 2024-09-05